+++ packages/core-js/actual/json/is-raw-json.js
@@ -1,0 +1,4 @@
+require('../../modules/esnext.json.is-raw-json');
+var path = require('../../internals/path');
+
+module.exports = path.JSON.isRawJSON;