+++ packages/core-js/actual/json/parse.js
@@ -1,0 +1,5 @@
+require('../../modules/es.object.keys');
+require('../../modules/esnext.json.parse');
+var path = require('../../internals/path');
+
+module.exports = path.JSON.parse;