+++ packages/core-js/actual/json/raw-json.js
@@ -1,0 +1,6 @@
+require('../../modules/es.object.create');
+require('../../modules/es.object.freeze');
+require('../../modules/esnext.json.raw-json');
+var path = require('../../internals/path');
+
+module.exports = path.JSON.rawJSON;