+++ packages/core-js/full/json/parse.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/json/parse');
+
+module.exports = parent;