+++ packages/core-js/full/json/raw-json.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/json/raw-json');
+
+module.exports = parent;