+++ packages/core-js/proposals/json-parse-with-source.js
@@ -1,0 +1,4 @@
+// https://github.com/tc39/proposal-json-parse-with-source
+require('../modules/esnext.json.is-raw-json');
+require('../modules/esnext.json.parse');
+require('../modules/esnext.json.raw-json');