+++ packages/core-js/stage/3.js
@@ -4,6 +4,7 @@ require('../proposals/array-from-async-stage-2');
 require('../proposals/array-grouping-stage-3-2');
 require('../proposals/explicit-resource-management');
 require('../proposals/iterator-helpers-stage-3-2');
+require('../proposals/json-parse-with-source');
 require('../proposals/set-methods-v2');
 require('../proposals/well-formed-unicode-strings');
 // TODO: Obsolete versions, remove from `core-js@4`